Closes: #1 Fixed validations.py script so that usernames must begin with a letter.
Before the change, validations.py was checking if the first character of the
username was a number, but failed to account for if the first character was a non-alpha
character such as a period (.) or an underscore (_).
The script was changed so that it checks using the isalpha() function to see if the first
character is an alphabetical characters, accounting for other non alphabetical characters.