Closes: #1
Updated validations.py python script.

Regular expression modified to ensure that the username does not start
with a number, underscore nor a dot, along this change, the line "return True"
has been added to dispatch a well formatted username.

Also, removed the line that checked if the username starts with a number
since the new Regular Expression handles this.