Bug fix: "print" erased at the end of the file, comments added
Below in the file was the "print" command, which allows us to test the file from the console. They were needed to test that the new function works, but they are not needed in the working file.

Test usernames, which are used to test the functionality of the new "first letter of the username" check, which can only be a letter, are added in the function comments in the code.

A description of the regex pattern has also been added to make it easy to read what affects what and how it can be easily modified.