Adding ValueError output to the screen + moved check pattern
After adding a check that the first letter of the user name should be only a letter, it was necessary to refine the name testing to visually display an error in case of violation of this rule (the beginning of the file).

After the check with showing an error (the first character of the name is not a letter) was added, it was necessary to transfer the "pattern" of the regex check from the place "if" (middle of the file) above to the place where the error was output (beginning of the file).