Closes: #1
To fix this issue, we can modify the regular expression to ensure
that the username doesn't start with a dot or underscore. We can
use the caret (^) to indicate that the pattern should start at the
beginning of the string and then specify that it should be followed
by a letter or number ([a-z0-9]).